Revert "   - Remove all in-memory repository implementations and test files"
This reverts commit 06cb155a0c2ccc542ad5fce0c6a639f9a4fad62d.

# Conflicts:
#	README.md
#	cmd/api/main.go
#	internal/handler/transaction/transaction.go
#	internal/idempotency/idempotency.go
#	internal/idempotency/postgres.go
#	internal/service/transaction/transaction.go
#	migrations/001_init.sql

+++ internal/model/transaction.go
@@ -21,7 +21,7 @@ type Transaction struct {
 	TransactionID   int64     `json:"transaction_id"`
 	AccountID       int64     `json:"account_id"`
 	OperationTypeID int64     `json:"operation_type_id"`
-	Amount          int64     `json:"amount"`
+	Amount          float64   `json:"amount"`
 	Type            string    `json:"type"`
 	EventDate       time.Time `json:"event_date"`
 	CreatedAt       time.Time `json:"created_at"`
